// risk-portal-ui/src/types/RegistrationData.d.ts
export interface RegistrationData {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  role?: string;
  vendor?: string;
  gender?: string;
  dateOfBirth?: string;
  phoneNumber?: string;
  companyName?: string;
  website?: string;
  size?: string;
  industry?: string;
  countryCode?: string;
  password: string;
  whatsappId?: string;
  subjectId?: string;
  username: string;
  position?: string;
}